Raycast & Mouse Lock
Set up a raycast system to detect scannable objects.
NOTE: Still undecided what part of the player it should be on, e.g head or camera

Set up the mouse lock script so that the mouse isnt visible on screen

Set up functionality for cursor that changes based on if it is looking at a scannable object or not.